refactor: moving typeform button post
