+++ src/components/study/card/StudyPostsCard.tsx
@@ -1,14 +1,12 @@
-import LinkButton from '@/components/base/LinkButton'
-import GlossaryModal from '@/components/modals/GlossaryModal'
 import { FullStudy } from '@/db/study'
-import { hasAccessToEmissionSourceValidation, hasAccessToPostTypeform } from '@/services/permissions/environment'
+import { hasAccessToEmissionSourceValidation } from '@/services/permissions/environment'
 import { Post, subPostsByPost } from '@/services/posts'
 import { withInfobulle } from '@/utils/post'
 import HelpOutlineIcon from '@mui/icons-material/HelpOutline'
 import { Environment } from '@prisma/client'
 import classNames from 'classnames'
 import { useTranslations } from 'next-intl'
-import { Dispatch, SetStateAction, useMemo, useState } from 'react'
+import { Dispatch, SetStateAction } from 'react'
 import PostIcon from '../infography/icons/PostIcon'
 import SelectStudySite from '../site/SelectStudySite'
 import styles from './StudyPostsCard.module.css'
@@ -26,8 +24,6 @@ const StudyPostsCard = ({ study, post, studySite, setSite, setGlossary, environm
   const t = useTranslations('study')
   const tPost = useTranslations('emissionFactors.post')
 
-  const [glossaryTypeform, setGlossaryTypeform] = useState('')
-
   const emissionSources = study.emissionSources.filter(
     (emissionSource) =>
       subPostsByPost[post].includes(emissionSource.subPost) && emissionSource.studySite.id === studySite,
@@ -35,31 +31,8 @@ const StudyPostsCard = ({ study, post, studySite, setSite, setGlossary, environm
   const validated = emissionSources.filter((emissionSource) => emissionSource.validated).length
   const percent = emissionSources.length ? Math.floor((validated / emissionSources.length) * 100) : 0
 
-  const showTypeformLink = useMemo(() => {
-    const typeformPosts: Post[] = [Post.DeplacementsDePersonne, Post.TransportDeMarchandises]
-    return (
-      process.env.NEXT_PUBLIC_TYPEFORM_DEPLACEMENTS_LINK &&
-      hasAccessToPostTypeform(environment) &&
-      typeformPosts.includes(post as Post)
-    )
-  }, [post, environment])
-
   return (
     <div className={classNames(styles.card, 'flex-col px1')}>
-      {showTypeformLink && (
-        <div className="justify-end align-center">
-          <LinkButton href={process.env.NEXT_PUBLIC_TYPEFORM_DEPLACEMENTS_LINK} rel="noreferrer noopener">
-            {tPost('seeTypeform')}
-          </LinkButton>
-          <HelpOutlineIcon
-            color="secondary"
-            className="pointer ml-2"
-            onClick={() => setGlossaryTypeform(tPost('seeTypeformDescription'))}
-            aria-label={tPost('seeTypeform')}
-            titleAccess={tPost('seeTypeform')}
-          />
-        </div>
-      )}
       <div className="justify-end align-center">
         <SelectStudySite
           sites={study.sites}
@@ -105,16 +78,6 @@ const StudyPostsCard = ({ study, post, studySite, setSite, setGlossary, environm
           </div>
         )}
       </div>
-      {glossaryTypeform && (
-        <GlossaryModal
-          glossary={'seeTypeform'}
-          label="typeform-glossary"
-          t={tPost}
-          onClose={() => setGlossaryTypeform('')}
-        >
-          {glossaryTypeform}
-        </GlossaryModal>
-      )}
     </div>
   )
 }
